launch-qemu: gdb: do not kill consoles on exit
In GDB mode, when gdb run hook executes gdb_run command, the script
setups up consoles and exits (as opposed to blocking on the qemu
invocation command in non-gdb mode), and the finish hook gets called. We
must not let the finish hook cleanup the consoles, because the whole
point of running gdb_run was to setup the consoles and exit, but leave
the console processes running asynchronously.